Fix build warnings and backend project filter
- C4099: forward-declare ibSourceDescription as `struct` (it is defined as a
  struct) in visualEditor.h.
- C4305/C4309: the Data and metadata module units register their member tags with
  a metatype CLSID that rides the (long) prop slot; narrow it explicitly with
  static_cast. Only the tag identity matters -- GetPropVal dispatches by ordinal
  and re-supplies the clsid itself -- so the narrowing is intentional, not a bug.
- backend.vcxproj.filters: file valueDynamicList.{cpp,h} under system\value. They
  were in the .vcxproj (compiled) but absent from the filters, so they floated at
  the backend project root in Solution Explorer after the earlier move.

Debug|x86 drops from 53 warnings to 2 (the residual C4005 macro-redefinitions are
pre-existing and out of scope).

+++ src/engine/backend/moduleManager/moduleManagerDataUnit.cpp
@@ -76,15 +76,17 @@ void ibValueModuleManager::ibValueDataUnit::FillMembers(ibMemberTable& helper) c
 
 	// The QUERYABLE kinds only — mirrors the query language's source coverage
 	// (queryableFactory): data-reference records, registers, constants.
-	helper.AppendProp("Constants", true, false, g_metaConstantCLSID);
-	helper.AppendProp("Catalogs", true, false, g_metaCatalogCLSID);
-	helper.AppendProp("Documents", true, false, g_metaDocumentCLSID);
-	helper.AppendProp("Enumerations", true, false, g_metaEnumerationCLSID);
-	helper.AppendProp("InformationRegisters", true, false, g_metaInformationRegisterCLSID);
-	helper.AppendProp("AccumulationRegisters", true, false, g_metaAccumulationRegisterCLSID);
-	helper.AppendProp("ChartsOfCharacteristicTypes", true, false, g_metaChartOfCharacteristicTypesCLSID);
-	helper.AppendProp("ChartsOfAccounts", true, false, g_metaChartOfAccountsCLSID);
-	helper.AppendProp("AccountingRegisters", true, false, g_metaAccountingRegisterCLSID);
+	// clsid narrowed into the (long) tag slot on purpose — GetPropVal keys by ordinal and re-supplies
+	// the clsid, so only the tag identity matters; static_cast makes the u64->long explicit (no C4305/C4309).
+	helper.AppendProp("Constants", true, false, static_cast<long>(g_metaConstantCLSID));
+	helper.AppendProp("Catalogs", true, false, static_cast<long>(g_metaCatalogCLSID));
+	helper.AppendProp("Documents", true, false, static_cast<long>(g_metaDocumentCLSID));
+	helper.AppendProp("Enumerations", true, false, static_cast<long>(g_metaEnumerationCLSID));
+	helper.AppendProp("InformationRegisters", true, false, static_cast<long>(g_metaInformationRegisterCLSID));
+	helper.AppendProp("AccumulationRegisters", true, false, static_cast<long>(g_metaAccumulationRegisterCLSID));
+	helper.AppendProp("ChartsOfCharacteristicTypes", true, false, static_cast<long>(g_metaChartOfCharacteristicTypesCLSID));
+	helper.AppendProp("ChartsOfAccounts", true, false, static_cast<long>(g_metaChartOfAccountsCLSID));
+	helper.AppendProp("AccountingRegisters", true, false, static_cast<long>(g_metaAccountingRegisterCLSID));
 }
 
 bool ibValueModuleManager::ibValueDataUnit::SetPropVal(const long lPropNum, const ibValue& varPropVal)
